+++ R/ard_effectsize_hedges_g.R
@@ -30,7 +30,7 @@
 #' The data are then passed as
 #' `effectsize::hedges_g(x = data_wide[[<by level 1>]], y = data_wide[[<by level 2>]], paired = TRUE, ...)`.
 #'
-#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("effectsize", "parameters"), reference_pkg = "cardx"))
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("effectsize", "parameters")))
 #' cards::ADSL |>
 #'   dplyr::filter(ARM %in% c("Placebo", "Xanomeline High Dose")) |>
 #'   ard_effectsize_hedges_g(by = ARM, variables = AGE)
@@ -52,7 +52,7 @@ ard_effectsize_hedges_g <- function(data, by, variables, conf.level = 0.95, ...)
   set_cli_abort_call()
 
   # check installed packages ---------------------------------------------------
-  check_pkg_installed(c("effectsize", "parameters"), reference_pkg = "cardx")
+  check_pkg_installed(c("effectsize", "parameters"))
 
   # check/process inputs -------------------------------------------------------
   check_not_missing(data)
@@ -101,7 +101,7 @@ ard_effectsize_paired_hedges_g <- function(data, by, variables, id, conf.level =
   set_cli_abort_call()
 
   # check installed packages ---------------------------------------------------
-  check_pkg_installed(c("effectsize", "parameters"), reference_pkg = "cardx")
+  check_pkg_installed(c("effectsize", "parameters"))
 
   # check/process inputs -------------------------------------------------------
   check_not_missing(data)
@@ -158,7 +158,7 @@ ard_effectsize_paired_hedges_g <- function(data, by, variables, id, conf.level =
 #'
 #' @return ARD data frame
 #' @keywords internal
-#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("effectsize", "parameters"), reference_pkg = "cardx"))
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("effectsize", "parameters")))
 #' cardx:::.format_hedges_g_results(
 #'   by = "ARM",
 #'   variable = "AGE",