+++ R/ard_stats_mcnemar_test.R
@@ -25,7 +25,7 @@
 #' The data is passed as `stats::mcnemar.test(x = data[[variable]], y = data[[by]], ...)`.
 #' Please use `table(x = data[[variable]], y = data[[by]])` to check the contingency table.
 #'
-#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "broom", reference_pkg = "cardx"))
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "broom"))
 #' cards::ADSL |>
 #'   ard_stats_mcnemar_test(by = "SEX", variables = "EFFFL")
 #'
@@ -47,7 +47,7 @@ ard_stats_mcnemar_test <- function(data, by, variables, ...) {
   set_cli_abort_call()
 
   # check installed packages ---------------------------------------------------
-  check_pkg_installed("broom", reference_pkg = "cardx")
+  check_pkg_installed("broom")
 
   # check/process inputs -------------------------------------------------------
   check_not_missing(data)
@@ -88,7 +88,7 @@ ard_stats_mcnemar_test_long <- function(data, by, variables, id, ...) {
   set_cli_abort_call()
 
   # check installed packages ---------------------------------------------------
-  check_pkg_installed("broom", reference_pkg = "cardx")
+  check_pkg_installed("broom")
 
   # check/process inputs -------------------------------------------------------
   check_not_missing(data)
@@ -138,7 +138,7 @@ ard_stats_mcnemar_test_long <- function(data, by, variables, id, ...) {
 #'
 #' @return ARD data frame
 #'
-#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "broom", reference_pkg = "cardx"))
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "broom"))
 #' cardx:::.format_mcnemartest_results(
 #'   by = "ARM",
 #'   variable = "AGE",