+++ tests/testthat/test-ard_continuous.survey.design.R
@@ -1,4 +1,4 @@
-skip_if_not(is_pkg_installed("survey", reference_pkg = "cardx"))
+skip_if_not(is_pkg_installed("survey"))
 
 test_that("unstratified ard_continuous.survey.design() works", {
   data(api, package = "survey")
@@ -517,4 +517,31 @@ test_that("ard_continuous.survey.design() follows ard structure", {
     ard_continuous(dclus1, variables = c(cds, stype), by = snum) |>
       cards::check_ard_structure(method = FALSE)
   )
+})
+
+test_that("ard_continuous.survey.design() original types are retained", {
+  data(api, package = "survey")
+  dclus1 <-
+    survey::svydesign(
+      id = ~dnum,
+      weights = ~pw,
+      data = apiclus1 |> dplyr::mutate(sch.wide_int = as.integer(sch.wide), sch.wide_dbl = as.numeric(sch.wide)),
+      fpc = ~fpc
+    )
+
+  # factors and integer check
+  expect_silent(
+    ard <-
+      ard_continuous(
+        data = dclus1,
+        variables = c(api00, api99),
+        by = c(stype, sch.wide_int, sch.wide_dbl)
+      )
+  )
+  expect_equal(
+    unlist(ard$group1_level) |> levels(),
+    levels(apiclus1$stype)
+  )
+  expect_true(unlist(ard$group2_level) |> is.integer())
+  expect_true(unlist(ard$group3_level) |> is.numeric())
 })